set OS_ env and keep running the processor for long time

+++ openstack/scripts/process.sh
@@ -3,26 +3,47 @@ THISDIR=$(dirname $0)
 REQ="${1}/request"
 RES="${1}/response"
 
-touch ${REQ}/status
-for r in $(ls -d ${REQ}/REQ-* 2>/dev/null | grep '/REQ-[a-zA-Z][a-zA-Z0-9_-]*$' || true) ; do
-  if [ -f $r ] ; then
-    rf=$(basename $r)
-    ERR=0
-    cmd=$(echo $rf | sed 's|^REQ-||;s|-.*||')
-    cmd2run=${THISDIR}/cmds/${cmd}.sh
-    if [ -f ${cmd2run} ] ; then
-      ${cmd2run} $r > ${RES}/${rf}.tmp 2>&1 || ERR=1
-    else
-      echo "ERROR: Invalid request $cmd" > ${RES}/${rf}.tmp
-      ERR=1
-    fi
-    rm -rf $r
-    echo "EXIT:$ERR" >> ${RES}/${rf}.tmp
-    mv ${RES}/${rf}.tmp ${RES}/${rf}
+LOCK="${REQ}/lock"
+if [ -f ${LOCK} ] ; then
+  let age=$(date +%s)-$(stat -c %Y ${LOCK})
+  if [ $age -gt 600 ] ; then
+    rm -f ${LOCK}
   else
-    rm -rf $r
+    exit 0
   fi
+fi
+XID="$(hostname -s):$$:"
+echo "${XID}" > ${LOCK}
+sleep 5
+let STIME=$(date +%s)+7200
+echo "$(date): ${XID}" >> ${REQ}/status
+while [ $(date +%s) -lt ${STIME} ] ; do
+  if [ $((grep "^${XID}" ${LOCK} 2>/dev/null || true) | wc -l) -eq 0 ] ; then
+    exit 0
+  fi
+  touch ${LOCK}
+  for r in $(ls -d ${REQ}/REQ-* 2>/dev/null | grep '/REQ-[a-zA-Z][a-zA-Z0-9_-]*$' || true) ; do
+    if [ -f $r ] ; then
+      rf=$(basename $r)
+      ERR=0
+      cmd=$(echo $rf | sed 's|^REQ-||;s|-.*||')
+      cmd2run=${THISDIR}/cmds/${cmd}.sh
+      if [ -f ${cmd2run} ] ; then
+        ${cmd2run} $r > ${RES}/${rf}.tmp 2>&1 || ERR=1
+      else
+        echo "ERROR: Invalid request $cmd" > ${RES}/${rf}.tmp
+        ERR=1
+      fi
+      rm -rf $r
+      echo "EXIT:$ERR" >> ${RES}/${rf}.tmp
+      mv ${RES}/${rf}.tmp ${RES}/${rf}
+    else
+      rm -rf $r
+    fi
+    touch ${LOCK}
+  done
+  sleep 10
 done
-
+rm -f ${LOCK}
 find ${REQ} -mindepth 1 -maxdepth 1 -mmin +59 | xargs  --no-run-if-empty rm -rf 
 find ${RES} -mindepth 1 -maxdepth 1 -mmin +59 | xargs  --no-run-if-empty rm -rf
